+++ src/com/android/server/telecom/RespondViaSmsManager.java
@@ -144,10 +144,10 @@ public class RespondViaSmsManager extends CallsManagerListenerBase {
     @Override
     public void onIncomingCallRejected(Call call, boolean rejectWithMessage, String textMessage) {
         if (call != null && rejectWithMessage) {
-            long subId = SubscriptionManager.getDefaultSubId();
+            int subId = SubscriptionManager.getDefaultSubId();
             if (call.getTargetPhoneAccount() != null) {
                 try {
-                    subId = Long.parseLong(call.getTargetPhoneAccount().getId());
+                    subId = Integer.parseInt(call.getTargetPhoneAccount().getId());
                 } catch (NumberFormatException e) {
                     Log.e(RespondViaSmsManager.this, e , "Exception e ");
                 }
