+++ src/com/android/server/telecom/TelecomService.java
@@ -623,7 +623,7 @@ public class TelecomService extends Service {
             long token = Binder.clearCallingIdentity();
             String retval = "content://icc/adn/";
             try {
-                long subId = mPhoneAccountRegistrar.getSubscriptionIdForPhoneAccount(accountHandle);
+                int subId = mPhoneAccountRegistrar.getSubscriptionIdForPhoneAccount(accountHandle);
                 retval = retval + "subId/" + subId;
             } finally {
                 Binder.restoreCallingIdentity(token);
@@ -702,10 +702,10 @@ public class TelecomService extends Service {
         }
 
         @Override
-        public long getActiveSubscription() { return 0; }
+        public int getActiveSubscription() { return 0; }
 
         @Override
-        public void switchToOtherActiveSub(long subId) { }
+        public void switchToOtherActiveSub(int subId) { }
 
         /**
          * Dumps the current state of the TelecomService.  Used when generating problem reports.
